pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "advulture"
version = "0.1.0"
description = "Active Directory Vulnerability Intelligence — Empirical Posture Analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "ADVulture Project" }]
keywords = ["active-directory", "security", "graph-neural-network", "markov", "posture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "ldap3>=2.9",
    "impacket>=0.12.0",
    "python-evtx>=0.8.0",
    "torch>=2.2.0",
    "torch_geometric>=2.5.0",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "duckdb>=0.10.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.0",
    "rich>=13.7.0",
    "streamlit>=1.32.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "msgraph-sdk>=1.3.0",
    "azure-identity>=1.16.0",
    "azure-monitor-query>=1.3.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "ipython>=8.22.0",
]
win = [
    "pywin32>=306; sys_platform == 'win32'",
    "wmi>=1.5.1; sys_platform == 'win32'",
]

[project.scripts]
advulture = "advulture.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["advulture*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
